[I25-178] refactor : group router를 통한 경로에 따른 Layout 사용 로직 변경
