+++ src/app/(box-layout)/editor/page.tsx
