+++ src/app/(box-layout)/portfolio/SearchTab.tsx
@@ -2,11 +2,11 @@
 
 import Link from 'next/link';
 import { useState } from 'react';
-import PortfolioCard from '../components/card/portfolio/PortfolioCard';
+import PortfolioCard from '@/app/components/card/portfolio/PortfolioCard';
 import { PortfolioData } from './types';
-import Inputs from '../components/modal/inputs/SingleInput';
-import Checkbox from '../components/modal/inputs/Checkbox';
-import { Body } from '../components/system/text';
+import Inputs from '@/app/components/modal/inputs/SingleInput';
+import Checkbox from '@/app/components/modal/inputs/Checkbox';
+import { Body } from '@/app/components/system/text';
 import { Tables } from '@/utils/supabase/database.types';
 
 const JOB_SEEKING_STATUS = ['구직 중', 'jobseeking'];
