+++ src/app/(box-layout)/portfolio/[profileName]/staticParamsGenerator.ts
