+++ src/app/(box-layout)/team/[teamName]/[projectName]/page.tsx
